+++ src/irvine/oeis/a054/A054625.java
@@ -15,6 +15,6 @@ public class A054625 extends Sequence0 {
 
   @Override
   public Z next() {
-    return ++mN == 0 ? Z.ONE : Integers.SINGLETON.sum(1, mN, k -> Z.SIX.pow(Functions.GCD.l(mN, (long) k))).divide(mN);
+    return ++mN == 0 ? Z.ONE : Integers.SINGLETON.sum(1, mN, k -> Z.SIX.pow(Functions.GCD.l(mN, k))).divide(mN);
   }
 }
