+++ src/irvine/oeis/a349/A349783.java
@@ -14,6 +14,6 @@ public class A349783 extends LambdaSequence {
 
   /** Construct the sequence. */
   public A349783() {
-    super(0, n -> Integers.SINGLETON.sum(0, n, j -> Functions.STIRLING1.z(2L * n, (long) j).abs()));
+    super(0, n -> Integers.SINGLETON.sum(0, n, j -> Functions.STIRLING1.z(2L * n, j).abs()));
   }
 }
